First commit.
This represents a first attempt at nailing the top part as a
navbar. That is, I wanted to divide the hero section into a navbar,
and a hero-section proper.

However, the proposed design itself doesn't suggest the
existence of a separate navbar. Hence, I want to try something
different: use only a top hero section, which is then vertically
divided in two.