Establish basic header content and styling.
So far so good. It looks like we'll be sticking with our current
approach moving forward.

I was sweating over how to get the "margined" effect, but for now I'm
resorting to a simple application of horizontal padding. I'm pretty
sure that's how you'd do it anyway.

I eliminated the actual image I was using, since its dimensions were
ruining the layout (or at least my ability to visualize it).

I'm trying to use a specific system right now when I write my
CSS. Basically, I try to apply multiple classes on a given HTML
element, to make any one class less complicated. I like to think of a
given class selector as being part of one of these chains; and so,
when it isn't, I try to use a more exact specifier.