Tweak CSS.
After pushing my work to GitHub and GitHub Pages, the page all of a
sudden didn't look the same as I was seeing it hosted locally. The
problem purportedly had to do with browser cache, and so I tried
loading the site in Firefox, and lo, the same mistakes appeared.

This commit attempts to fix those mistakes. We can only be sure,
though, after pushing.